audio: Fix wired headset routing issue
When quad-spkr is enabled and headphone detection
received, route is not occurring to headphone because it satisfy
is_quad_spkr check and enables quad speaker.
Add check for sound device is speaker or not which ensures only
force routing happens for speaker device.
Also resend mixer coefficients on every device re-route.

Change-Id: Ib308ba3eaadd3095954e4acc7482404070cc0d5d